requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.8
django-appconf==1.0.3
django-bootstrap-form==3.4
django-debug-toolbar==2.1
django-forms-bootstrap==3.1.0
django-tables2==2.2.0
django-taggit==1.1.0
django-test-plus==1.3.1
django-user-accounts==2.1.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
mccabe==0.6.1
Pillow==6.2.1
pinax-theme-bootstrap==8.0.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
